clang-tidy sweep + tooling: type widening, safe string ops, parallel driver
Source fixes:
- sortOrder: int8_t -> uint8_t in SomfyShade/SomfyGroup/SomfyRoom; the three
  getMax*Order accumulators widen to int16_t internally and narrow to int8_t
  on return so the public API is unchanged.
- repeat (HTTP command handlers): int8_t -> int16_t to avoid narrowing from
  atoi().
- Web::handleBackup loop index: uint8_t -> size_t (matches strlen).
- WebOTA / Web token build: strcpy/strcat -> strlcpy/strlcat.
- Web::isAuthenticated: collapse two return-true branches into a single
  authNotRequired guard; silences bugprone-branch-clone and drops the
  brittle == 0x01 mask compare.
- CC1101 SpiStart: explicit static_cast<int8_t> on pin args.
- CC1101 setPA: initialise local `int a = 0` so the PA table isn't
  programmed with stack garbage when MHz falls outside supported bands.

Tooling:
- run-clang-tidy.py now drives the parallel `run-clang-tidy` from the
  newest esp-clang on disk, forces -use-color only on a TTY, and skips
  vendor subtrees (ArduinoJson, CC1101, PubSubClient, WebSockets,
  esp-homekit-sdk) at the compile-DB step.
- bump_version.sh: pre-bump reminder for host unit tests and clang-tidy,
  with a y/N confirmation before any file is written.

+++ main/somfy/SomfyShadeController.cpp
@@ -221,23 +221,23 @@ uint8_t SomfyShadeController::getNextShadeId() {
 }
 
 int8_t SomfyShadeController::getMaxShadeOrder() {
-  int8_t order = -1;
+  int16_t order = -1;
   for(uint8_t i = 0; i < SOMFY_MAX_SHADES; i++) {
     SomfyShade *shade = &this->shades[i];
     if(shade->getShadeId() == 255) continue;
     if(order < shade->sortOrder) order = shade->sortOrder;
   }
-  return order;
+  return static_cast<int8_t>(order);
 }
 
 int8_t SomfyShadeController::getMaxGroupOrder() {
-  int8_t order = -1;
+  int16_t order = -1;
   for(uint8_t i = 0; i < SOMFY_MAX_GROUPS; i++) {
     SomfyGroup *group = &this->groups[i];
     if(group->getGroupId() == 255) continue;
     if(order < group->sortOrder) order = group->sortOrder;
   }
-  return order;
+  return static_cast<int8_t>(order);
 }
 
 uint8_t SomfyShadeController::getNextGroupId() {
@@ -281,13 +281,13 @@ uint8_t SomfyShadeController::getNextRoomId() {
 }
 
 int8_t SomfyShadeController::getMaxRoomOrder() {
-  int8_t order = -1;
+  int16_t order = -1;
   for(uint8_t i = 0; i < SOMFY_MAX_ROOMS; i++) {
     SomfyRoom *room = &this->rooms[i];
     if(room->roomId == 0) continue;
     if(order < room->sortOrder) order = room->sortOrder;
   }
-  return order;
+  return static_cast<int8_t>(order);
 }
 
 uint8_t SomfyShadeController::repeaterCount() {
